adds card component

+++ 01-starting-setup/src/ExpenseDate.js
@@ -2,7 +2,7 @@ import './ExpenseDate.css';
 
 function ExpenseDate(props) {
     const day = props.date.toLocaleString('en-GB',{day:'2-digit'});
-    const month =props.date.toLocaleString('en-GB',{month:'short'});
+    const month =props.date.toLocaleString('en-GB',{month:'long'});
     const year = props.date.getFullYear();
 
     return (
